server/requirements.txt
fastapi
uvicorn[standard]
langchain
httpx
python-dotenv
langchain-google-genai
langchain-google-community[calendar]
notion-client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai